[S07P12D102-290] FIX : 오류 메시지 추가

+++ BE/CommonProject/src/main/java/com/project/common/service/Group/GroupService.java
@@ -68,6 +68,8 @@ public class GroupService{
 	//모임 삭제
 	public String deleteGroup(int groupSeq){
 		GroupEntity group = groupRepository.findByGroupSeq(groupSeq);
+		if(group==null)
+			return "Fail - No Group";
 		for(GroupMemberEntity entity : group.getMembers()) {
 			groupMemberRepository.deleteByMemberSeq(entity.getMemberSeq());
 		}
